chore: move form-step folder
