+++ ui/src/views/mint/nfa-step/form-step/fields/app-description-field.tsx
@@ -1,5 +1,5 @@
 import { Form } from '@/components';
-import { Mint } from '../../mint.context';
+import { Mint } from '../../../mint.context';
 
 export const AppDescriptionField = () => {
   const { appDescription, setAppDescription } = Mint.useContext();
