+++ ui/src/views/mint/nfa-step/form-step/fields/index.ts
@@ -2,4 +2,3 @@ export * from './logo-field';
 export * from './app-name-field';
 export * from './app-description-field';
 export * from './ens-domain-field';
-export * from './verify-nfa-field';
