+++ ui/src/views/mint/nfa-step/form-step/fields/logo-field/index.ts
