+++ ui/src/views/mint/nfa-step/form-step/fields/logo-field/logo-field.tsx
@@ -1,6 +1,6 @@
 import { Flex, Form } from '@/components';
 import { useState } from 'react';
-import { Mint } from '../../../mint.context';
+import { Mint } from '../../../../mint.context';
 import { fileToBase64, validateFileSize } from '../../form.utils';
 import { ColorPicker } from './color-picker';
