+++ ui/src/views/mint/nfa-step/form-step/form.utils.ts
