+++ ui/src/views/mint/nfa-step/form-step/index.tsx
